refactor: Card 객체 설계 수정
